Refactor: move project configurable iface to core
rename to ProjectConfigurable
